refactor: Rename context menu command files for consistency and clarity
